feat(examples): fold composition examples into the guide (phase 3)
Third phase-3 cluster. The 6 standalone example pages for context, callbacks,
virtualize, keyed lists, drag & drop, and error boundaries are removed; each is
now an inline live demo in the Composition guide.

- Register the 10 composition demos in DemoRegistry.
- Add markers throughout docs/composition.md, plus new "Keyed lists" and
  "Error boundaries" sections; replace the runnable-demo pointers to deleted pages.
- Promote the keyed-lists interactive widget (keys on/off toggle, reorder,
  focus-preserving rows with #kl-list/#kl-reverse/.kl-note) from the page into
  KeyedListsReorderDemo so it stays runnable inline; retarget KeyedListsPageTests
  and BoomPageTests to the demos.
- Delete the 6 *Page.cs; drop their ShowcaseLayout links, HomePage cards, and the
  Boom/Virtualize PageBaselineTests rows.
- Consolidate the E2E into a new TestCompositionGuideAsync that opens the guide
  once and drives all six demos in place (scoped locators).
- CHANGELOG updated.
