+++ docs/composition.md
@@ -145,6 +145,11 @@ It is cooperative: a handler that ignores the token (or runs unbounded synchrono
 force-aborted — that's a .NET reality, not a Rask limitation. In a lifecycle hook (no handler
 dispatch) the token is simply the component's lifetime token.
 
+A child raises an event through a plain delegate prop; the framework wraps it so the click re-renders
+the owning parent — no `StateHasChanged`:
+
+<!-- demo:callback-rating -->
+
 ---
 
 ## Context: provide / consume
@@ -186,7 +191,7 @@ it re-reads when the provider re-renders — *even through a render-cached inter
 that never re-renders itself. That is the whole point: `ThemeCard` above is cached after
 first paint, yet the `ThemeBadge` it nests still updates on every toggle.
 
-Runnable demo: [`samples/Rask.Example.Shared/Features/Context/ContextThemeDemo.cs`](../samples/Rask.Example.Shared/Features/Context/ContextThemeDemo.cs).
+<!-- demo:context-theme -->
 
 ---
 
@@ -229,7 +234,23 @@ ItemsProvider: async req =>
 Provider mode caches by index, marks rows `IsPlaceholder` while a page is in flight, and
 cancels + disposes superseded requests (and on unmount).
 
-Runnable demo: [`samples/Rask.Example.Shared/Features/Virtualize/VirtualizePage.cs`](../samples/Rask.Example.Shared/Features/Virtualize/VirtualizePage.cs).
+**Items mode** — a fixed in-memory list, windowed:
+
+<!-- demo:virtualize-items -->
+
+**Provider mode** — rows fetched on demand as they scroll into view:
+
+<!-- demo:virtualize-provider -->
+
+---
+
+## Keyed lists
+
+A `Key:` on a list item gives it a stable identity across renders, so a reorder **moves** the live
+DOM node (with its focus, caret, and uncommitted input) instead of detaching and re-creating it. This
+is the same reconciliation identity the diff uses everywhere — not a reactive prop.
+
+<!-- demo:keyed-lists-reorder -->
 
 ---
 
@@ -237,9 +258,31 @@ Runnable demo: [`samples/Rask.Example.Shared/Features/Virtualize/VirtualizePage.
 
 A headless drag-and-drop primitive lives in `Rask.Core/DragAndDrop`. It tracks the
 dragged item and the drop target and raises a callback when an item is dropped; you own
-the visuals. See the runnable demo:
-[`samples/Rask.Example.Shared/Features/DragDrop/DragDropPage.cs`](../samples/Rask.Example.Shared/Features/DragDrop/DragDropPage.cs)
-(+ `DragDropPage.css`).
+the visuals — a sortable list and a kanban board built on the same primitive:
+
+<!-- demo:drag-drop-sortable -->
+
+<!-- demo:drag-drop-kanban -->
+
+---
+
+## Error boundaries
+
+An error boundary catches an exception thrown by a descendant — during an event handler **or** during
+render — and shows a fallback instead of tearing down the whole app. The nearest boundary handles it;
+everything outside it (the navbar, the rest of the page) keeps running, and `Recover` restores the
+healthy subtree. Boundaries nest, so a local failure stays local.
+
+<!-- demo:boom-handler -->
+
+A render-time throw is rewound cleanly (the serializer discards the partial output) and caught exactly
+once:
+
+<!-- demo:boom-render -->
+
+Nested boundaries — the innermost one catches, leaving its siblings untouched:
+
+<!-- demo:boom-nested -->
 
 ---
