+++ samples/Rask.Example.Shared/Features/Home/HomePage.cs
@@ -23,21 +23,15 @@ public sealed class HomePage(Navigator nav) : Component
         ("Components", "bi-link-45deg", "Route + query params", "Bind URL segments and query.", "/users/42"),
         ("Components", "bi-compass", "Navigator", "Programmatic navigation.", "/navigator"),
         ("Components", "bi-arrow-repeat", "Lifecycle", "Mount, props-changed, rendered hooks.", "/lifecycle"),
-        ("Components", "bi-diagram-2", "Context", "Provide and consume without prop drilling.", "/context"),
-        ("Components", "bi-arrow-up-right-circle", "Callbacks", "Child→parent events, auto re-render.", "/callback"),
         ("Components", "bi-bell", "Toast", "Show, stack, dismiss & auto-hide — no JS.", Routes.ToastPage()),
         ("Components", "bi-bullseye", "Element refs", "Reach the live DOM from C#.", "/element-ref"),
         ("Components", "bi-x-circle", "Cancellation", "A token that fires on unmount.", "/cancellation"),
         ("Components", "bi-trash", "Disposal", "IDisposable / IAsyncDisposable cleanup.", "/disposal"),
         ("Components", "bi-mouse", "Events", "DOM event handlers.", "/events"),
-        ("Components", "bi-list-ol", "Virtualize", "Headless windowed lists.", "/virtualize"),
         ("Components", "bi-table", "Data table", "Sortable, paginated table.", "/table"),
         ("Components", "bi-list-nested", "Master-detail", "Collapsible rows with a nested datagrid.", "/master-detail"),
-        ("Components", "bi-key", "Keyed lists", "Stable identity for trusted diffs.", "/keyed-lists"),
-        ("Components", "bi-arrows-move", "Drag & drop", "Headless reordering primitive.", "/drag-drop"),
         ("Components", "bi-graph-up-arrow", "Live ticker", "Server-pushed live updates.", "/realtime/BTC"),
         ("Components", "bi-person-lock", "User & auth", "Gate UI on the current user.", "/user"),
-        ("Components", "bi-shield-exclamation", "Error boundary", "Catch render and lifecycle errors.", "/boom"),
 
         ("Styling", "bi-palette", "Scoped CSS", "Co-located, isolated component styles.", "/scoped-css"),
         ("Styling", "bi-link-45deg", "Asset loading", "Content-addressed scoped assets.", "/asset-loading"),
