+++ samples/Rask.Example.Shared/Shared/ShowcaseLayout.cs
@@ -32,8 +32,6 @@ public sealed class ShowcaseLayout(RouteState route, IEnumerable<ShowcaseNavEntr
         (Features.Routes.UserDetailPage("42"), "Route + query params", "bi-link-45deg", "Components", "/users"),
         (Features.Routes.NavigatorPage(), "Navigator", "bi-compass", "Components", null),
         (Features.Routes.LifecyclePage(), "Lifecycle", "bi-arrow-repeat", "Components", null),
-        (Features.Routes.ContextPage(), "Context", "bi-diagram-2", "Components", null),
-        (Features.Routes.CallbackPage(), "Callback", "bi-arrow-up-right-circle", "Components", null),
         (Features.Routes.ToastPage(), "Toast", "bi-bell", "Components", null),
         (Features.Routes.ElementRefPage(), "Element refs", "bi-bullseye", "Components", null),
         (Features.Routes.UserPage(), "User & auth", "bi-person-lock", "Components", null),
@@ -42,12 +40,8 @@ public sealed class ShowcaseLayout(RouteState route, IEnumerable<ShowcaseNavEntr
         (Features.Routes.CancellationPage(), "Cancellation", "bi-x-circle", "Components", null),
         (Features.Routes.DisposalPage(), "Disposal", "bi-trash", "Components", null),
         (Features.Routes.EventsPage(), "Events", "bi-mouse", "Components", null),
-        (Features.Routes.VirtualizePage(), "Virtualize", "bi-list-ol", "Components", null),
         (Features.Routes.TablePage(), "Data table", "bi-table", "Components", null),
         (Features.Routes.OrdersPage(), "Master-detail", "bi-list-nested", "Components", null),
-        (Features.Routes.KeyedListsPage(), "Keyed lists", "bi-key", "Components", null),
-        (Features.Routes.DragDropPage(), "Drag & drop", "bi-arrows-move", "Components", null),
-        (Features.Routes.BoomPage(), "Error boundary", "bi-shield-exclamation", "Components", null),
         (Features.Routes.SvgPage(), "SVG", "bi-vector-pen", "DSL", null),
         (Features.Routes.ScopedCssPage(), "Scoped CSS", "bi-palette", "Styling", null),
         (Features.Routes.AssetLoadingPage(), "Asset loading", "bi-link-45deg", "Styling", null),
